Expose ARP/NDP state/ttl to FBOSS CLI
Summary: Now that we have landed arp/ndp expiration, it would be nice to expose some of the details about neighbor entries over the cli for debugging purposes. This task is to add thrift endpoints to get that data and then also add support to our cli tool to display it.

Reviewed By: aeckert

Differential Revision: D3004734

fb-gh-sync-id: 4777868f8ec21ee5bdd4dc1e4f75a305177ca60b
fbshipit-source-id: 4777868f8ec21ee5bdd4dc1e4f75a305177ca60b