Removing "cli_common.py" module and merge into "commands.py"
Summary:
The code in cli_common.py was only used by PortDetailsCmd. It also needed to be refactored. This is only for the open source version. I will be cleaning up our inteneral version as well.
This effort is to help move us closer to being able to opens source this tool.

- Moved all the code into commands.py "PortDetailsCmd" class.
- Refactored the nested functions
- Aligned the left column of the errors/bps output of the command (it was unaligned)
- Added a space and more text to the specific port details output when you request one port
- Refactor "convert_bps"

Reviewed By: oridb

Differential Revision: D2902663

fb-gh-sync-id: ea5efb940cecc78affaf5240423789659bb94446
fbshipit-source-id: ea5efb940cecc78affaf5240423789659bb94446