Finish the output to ODS on SwSwitch & Main file. Have problem with BcmSwitch
Summary: Added two functions at facebook/SwSwitch.cpp to do the ods log for SwSwitch & Main. The enum is defined inside of Util.h. Currently we can see the log with the two keywords: fboss.ctrl.hw_initialized, fboss.ctrl.hw_configuration_done. But when I am trying to change the code into Util to make it easier for BcmSwitch to invoke (the codes are in comments), the compiler stop me for the reason some headers not found. So I now need some information to see how should I go through that problem.

Differential Revision: D2952256

fb-gh-sync-id: f98356c84fc3388638fdad0e88a64345b7819f93
fbshipit-source-id: f98356c84fc3388638fdad0e88a64345b7819f93