Fix race between TunManager::probe and TunManager::sync
Summary:
There was actually a race in our tun interface code. During SwSwitch::init
we ask TunManager to do a probe to figure out all the existing tun interfaces.
This probe is scheduled to happen in backgroundEventBase. The main thread then
continues on and immediately after init applies the initial configuration. Once
initial config is applied we ask TunManager to sync the interfaces that it
discovered via probe vs what came in in the config. However there is no guarantee
that the initial probe is done yet. If its not done, we would try to add a
already configured interface + plus source routes for its subnet and encounter
errors leading to a abort.

Why do we see this now ? D2952256 added some logging of ODS data. This logging
happens in backgroundEventBase and is scheduled prior to TunManager::probe. This
affected the timing a bit and hence we started seeing the errors. As is typical
of races, we don't see it on all systems and all binaries. So wedge_agent on
6-packs show this but not wedge_ctrl on RSWs.

Fix is to check in sync whether a probe was done or not. If not done ask for
it to be done right away. While at it moved this logging to after TunManager::startProbe
call so it gets scheduled earlier than the ODS logging code and in the common
case sync doesn't have to wait at all.

Reviewed By: aeckert

Differential Revision: D3073551

fb-gh-sync-id: 61ba8fd812445d3f2d0b635f8df957af682df9f0
fbshipit-source-id: 61ba8fd812445d3f2d0b635f8df957af682df9f0