fix ecmp expand during initialization
Summary:
There is a race (described well in the linked task) on init
with ecmp expand. Basically we may not correctly process a port flap
during initialization. When the port goes down during init, the
SwSwitch may not be ready and won't handle the link down event. Then
when the port comes back the SwSwitch thinks it has been up the entire
time and does not communicate anything to the hw. Ideally the hw could
add the entries back itself, but this leads to packet loss.

This diff changes the bool for whether an entry is pending to an enum
for the 'state' of an entry in the NeighborEntry SwitchState node. A
NeighborEntry can be in three different states: REACHABLE, PENDING or
UNVERIFIED. We already handled the first two before with the pending
bool. We only use UNVERIFIED on startup for entries we think are
reachable. As soon as we know this entry is reachable, we change the
state from UNVERIFIED to REACHABLE, which generates a state update
and gives the hw a chance to expand the ecmp group.

Differential Revision: D3031973

fb-gh-sync-id: 7d68e51a1d65352a9670dc5f79e13d34e48a73e8
fbshipit-source-id: 7d68e51a1d65352a9670dc5f79e13d34e48a73e8