Don't serialize/deserialize neighbor entry timers
Summary:
As title. These timers were added in D2004266, if we
have a agent running from before this revision it won't dump these
timers and then when we come around to deserialize them we crash.
Actually there is no need to dump these values to switch_state file.
Unless they are changed in config we just dump and then read back
the same constant values. If they are overridden in config, we will
apply the config in a bit to override them anyway.

Reviewed By: oridb

Differential Revision: D3095701

fb-gh-sync-id: f132315593472ee480a42531964aa5979efae92b
fbshipit-source-id: f132315593472ee480a42531964aa5979efae92b