Use port names in port status command and add speed
Summary:
This changes the port status command to use port names (where
available) instead of port id. It also no longer displays disabled
ports, which are more often than not the sub-ports of a larger port.

I also added speed and shortened some of the labels. Finally changed
Transceiver to no longer display "Not Present". We only print if the
transceiver is present, or we print "Unknown" if there was some other
error.

For more info on the motivation for these changes see:
https://fburl.com/229735737

Reviewed By: cooperlees

Differential Revision: D3081881

fb-gh-sync-id: dd86ad34a77cfe61f7ed134c1e01a254296bd9b7
fbshipit-source-id: dd86ad34a77cfe61f7ed134c1e01a254296bd9b7