Make NDP handling follow the spec more accurately
Summary:
This updates our NDP cache code to more closely mimic the NDP
RFC. The issue we observed is that some OOBs send two ndp neighbor
advertisements in response to an ndp neighbor solicitation. My logic
before was to treat any ndp packet we receive as solicited if we have
an outstanding ndp solicitation. This breaks down in this case and we
treat the second entry as 'unsolicited' and set the entry to stale.
This causes us to send out another solicitation as soon as we see more
traffic and we loop forever sending neighbor solicitations.

This changes the logic to instead read the flags in the packet for
whether the packet was solicited. I also now respect the override
flag properly and follow the rest of the rules for receiving ndp
neighbor solicitations and advertisements.

Differential Revision: D3119378

fb-gh-sync-id: 9383341aa19f452a543b916b81dc6f049119e99c
fbshipit-source-id: 9383341aa19f452a543b916b81dc6f049119e99c