Add new assembly and basic link detection
Summary:
This diff adds very rudimentary LED support to wedge100. All
that it does now is show blue if the port is up. My next diff will
make this more robust, being able to change to different
colors/compact the LEDs. The main thing to look at for this diff is the
assembly code. I more or less emulated what we do for wedge40. I did
remove the blinking functionality, though we can add that back in later
if need be.

Reviewed By: simpkins

Differential Revision: D3110113

fb-gh-sync-id: f2da4448812436e02f77b0941a4f62a191f693e3
fbshipit-source-id: f2da4448812436e02f77b0941a4f62a191f693e3