Add color based on lane configuration and add compact mode
Summary:
This adds logic to the Wedge100Port to do a few things:

1. Detect the lane mode for the corresponding port group.
2. Detect the lane mode for the neighbor port (above or below).
3. If both of these are in DUAL or SINGLE mode, we use what I call
   "compact" mode. This means we write the same values for both the
   top four LEDs and the bottom four LEDs so the button has no effect.

I also set the color scheme based on lane configuration.

SINGLE (1x40, 1x100) -> BLUE
DUAL (2x50) -> CYAN
QUAD (4x10, 4x25) -> GREEN

I can change this, but it has a few benefits. First as we transition
40G -> 100G and 10G -> 25G the LED colors will stay the same. We also
won't have a ton of different colors. The HW guys are more comfortable
with different colors for each speed, but I was thinking this could be
more useful.

There is still no RX/TX blinking activity. I may add that in a later diff,
as well as making the LED red on errors.

Reviewed By: tfangit

Differential Revision: D3110225

fb-gh-sync-id: 2a757e6262d37b26e7d142e752c272d79504fff1
fbshipit-source-id: 2a757e6262d37b26e7d142e752c272d79504fff1