Handle TTL/state being None.
Summary: When a switch whose Thrift interface does not support exposing ARP/NDP entries' TTL and state, a None value is returned for these fields. In these cases, the CLI will exit with an exception because of attempting to apply the division operator to the value None. To decouple the next CLI release from the next agent release, and because we have several hundred switches whose agents will not support retrieving these fields until a disruptive upgrade, this patch prints "NA" and "?" for the state and TTL fields when they are not available.

Reviewed By: aeckert

Differential Revision: D3160464

fb-gh-sync-id: 8b11a5008ad849e6d7e8eb45ba9eeb5917736854
fbshipit-source-id: 8b11a5008ad849e6d7e8eb45ba9eeb5917736854