sixpack: fboss cli: fix issue in port sort function
Summary:
fboss cli was recently changed to print information sorted by port names.
This was working fine for wedges, but fails for SixPack.  The reason turned
out to be a minor issue in port_sort_fn which was not returning a tuple
when port name is empty.

Note: the reason it did not fail on wedges is because on wedges even ports
that are in POWER_DOWN state are given names in the config, whereas in
sixpack config they are not.

Differential Revision: D3167509

fb-gh-sync-id: af7b5ac7213a6a4ff46fed2c204997fd616568a5
fbshipit-source-id: af7b5ac7213a6a4ff46fed2c204997fd616568a5