add support for logging route updates
Summary:
We want to be able to enable logging for route updates to a prefix.
Add a state observer class that can react to route deltas and call into
a helper class which synchronizes access to a radix tree to decide whether
to log or not. Dependent diff expose it over thrift.

Differential Revision: D3143021

fb-gh-sync-id: 0ab95eef54e11e3bf50272d103fe1c8110679895
fbshipit-source-id: 0ab95eef54e11e3bf50272d103fe1c8110679895