port group fixes for 6.4.10
Summary:
It looks like the bcm_port_subsidiary_ports_get call works
slightly differently in 6.4.10. Instead of just returning the first
lane as a controlling port, the third lane is also a controlling port
and no longer throws a BCM_ERROR_PORT error. To work around this, this
changes our logic to not try to create a port group for any port that
is already in a group.

Differential Revision: D3219724

fbshipit-source-id: f4a5865c215a98832530fb737395af5f74dbf01a