Have platforms define a list of supported lane speeds
Summary:
Previously, we only set the maximum lane speed for a platform
port and then determined how many lanes we need based on that. This
was buggy and we only really got away with it because we were always
using 10G channels.

The problem is that we need to know all possible channel speeds and
ensure that the speed of the port can be made up of 1, 2 or 4 lanes
of one of those speeds. To make this concrete:

100G port -> 4 x 25G lanes
50G port -> 2 x 25G lanes
40G port -> 4 x 10G lanes

This changes the code for flexports to determine the number of needed
lanes for a port based on the supported lane speeds, not just the max
speed. This will correctly handle the case above.

Differential Revision: D3219759

fbshipit-source-id: c5897fbd20f7f1a47597c06bf201fecdc0c60618