Host entries can now change since host routes are in the host tables too
Summary:
With the move to putting host routes in host tables
we could get into a situation where post warm boot we hear
of a host route that has different next hops than pre warm
boot. Since a lot of host routes point to VIPs, which tend
to churn this is very possible. I saw this occur quite a
few times when rolling in atn1c16 test bed. There are 2
ways to fix this
a) Relax the check so changed host entries just update existing
ones
b) Reconstruct routes along with host entries in initial warm boot
state. This gets played before we hear any routes from bgpd
so this should match the state agent had before exit and the
check would still hold.

I went with b) since we want this anyway for t9902379

(There is still some more testing that I want to do with this
and add some more checks to make it safer. But sending for
early review).

Reviewed By: aeckert

Differential Revision: D3246439

fbshipit-source-id: ae529adf099e6f58c9d5f10656eb9e441db51882