Add a scuba fib update logger so we can track all fib updates in one place
Summary:
As title. For FB builds we default to scuba logger.
This would be useful when tracking updates across multiple
systems (e.g. during a roll or while debugging). Use the
async offer api on scribe so we don't block on writes.
Added a try/catch as a paranoia measure even tho offer
should never throw. Tested with and w/o scribe running
on the box, which I shut down links to get route updates.

Differential Revision: D3239669

fbshipit-source-id: e1cbd4d343519b2c5fd6f575f5af202b2384e79b