set Wedge100 LED mode to 12-bit mode on startup
Summary:
This sets the wedge100 led mode to 12-bit when the agent
starts up. The default LED mode is rainbow mode for now. The benefit
of this is that accton can verify LEDs. Also rainbow mode is a clear
indication that the rack is not up yet.

This may change if we get Accton to load LED functionality on the box, but this can't hurt for now.

I added read/write calls to the BaseWedgeI2CBus class (and
WedgeI2CBusLock), these could be in the CP2112 class itself, but since
this has some higher level logic it seemed like it should go in the
BaseWedgeI2CBus class.

Reviewed By: tfangit

Differential Revision: D3199212

fbshipit-source-id: ada63dcfed67e204a85085d07a1e3bffbcf49f8f