Fix race on b/w NeighborCache and NeighborCacheEntries on destruction
Summary:
When NeighborCache was destructed it called
destroy on each of the entries from the base NeighborCache
destructor. NeighborCacheEntries hold pointer to their
neighbor cache and can call methods asynchrounous to this
destruction on the neighbor cache. If such a method was
called while the entries were being destroyed from the
base neighbor cache destructor, this method would be called
on the base neighbor cache class. If this method happened to
be pure virtual, we get a pure virtual function called exception
and crash.

To fix this, ensure that till entries are destroyed the associated
cache objects are still alive. To do that move the entry destruction
to another method (clearCache) that gets called by NeighborUpdater (which
owns the caches) before it actually destroys the caches.
Looking at code in isolation there is still a possible race b/w
NeighboCache::clearCache being called and the NeighborCache being
destroyed. In that if entries got created in b/w this time we
would have the same race as above. However the only way to create
entries in NeighborCache is via state NeighborUpdater or by receiving
a ARP/NDP reply. However this doesn't seem possible in our code as
a) clearCache is called from NeighborUpdater's destructor, so at
this time NeighborUpdater is no longer observing state updates.
b) We shut down packet receive call back before we destroy NeighborUpdater.

However for paranoia's sake make the pure virtual function not
be pure virtual and log a DFATAL instead.

Reviewed By: aeckert

Differential Revision: D3288511

fbshipit-source-id: 0091b2c353a9bcadbf3e8d7d4382ef71d5f04dec