use last ip address for interface routes when applying switch config
Summary:
While logging updates during fib sync, we found that we have
a spurious update when the switch has an interface configured with
more than one address. The interface route programmed while going
through the config is the first one we see. However, RouteUpdater,
which also has a method for adding interface routes
(addInterfaceAndLinkLocalRoutes) uses the last address in the configuration.
Make the two consistent, and add a test where we have several
addresses for one interface configured.

Differential Revision: D3296063

fbshipit-source-id: 4185edd04f089af9d185d791317aac243297a031