Make sure all state is written to disk before we write can_warm_boot marker
Summary:
We were writing the switch_state file to disk *after* writing
the can_warm_boot marker file. Till the time we only dumped switch state
but never relied on it on warm boot this sort of worked. However now that
we do rely on this file, make sure that all state is synced to disk
before we write the warm boot marker.

There was a interesting race we saw in t11191205, where BRCM code
crashed while manipulating interrupt handlers after unit had been
detached and all of BRCM state synced to disk. In this case everything
was in place to do a warm boot, however since we were writing the switch
state file after, we got a truncated switch state file and crashed on
next boot.

Differential Revision: D3296140

fbshipit-source-id: b1053c5e3293d8772641c814ed9e76decc16cd22