make neighbor cache clear a bit cleaner by using futures
Summary:
current logic is slightly convoluted but functional. This simplifies
it a bit to no longer create futures that will asynchronously register
asynchronous cancelTimeout calls onto the event base. Instead, just get
the future for the destroy() itself, and block on that before clearing.
In fact, blocking on the destroy() is not actually necessary, so we
can get rid of that as well, but to make the code functionally identical,
I figured I could leave it in for now.

Reviewed By: aeckert

Differential Revision: D3302303

fbshipit-source-id: f6ea761a16587b9cb0cae5d750db68356d004320