log all route updates until fib sync
Summary:
after we notify the state observers of the initial state delta
(against empty switch state), turn on logging for all routes. That way,
any routes that get updated by the initial fib sync will be logged. Disable
all logging after this is as it will be noisy and possibly expensive.

Differential Revision: D3272969

fbshipit-source-id: af5272e4afc061c582abf3a7b0db7d83d8bb0eec