Added BroadView Packet Trace support
Summary:
This diff introduces BroadView Packet Trace (PT) support to FBOSS. We added a getPacketTrace thrift call to FBOSS ctrl thrift. Internally, we route the thrift call through SwSwitch, HwSwitch, and to OpenNSL APIs. Also included in this diff is a Python client based on Scapy3k.

With PT, we can inject an arbitrary packet into the ASIC and observe how it's processed. One useful information is egress port out of an ECMP group.

The function is only available to Broadcom SDK > 6.4.6. The feature itself is only available on Tomohawk, not Trident2. Hence the thrift call will throw "feature not available" error on Trident2 platform.

Differential Revision: D3415989

fbshipit-source-id: 407968c0acc3b10b161c699a00a11b188f5af4a2