replace LockAndStatItem with new LockableStat class in fboss/agent directory
Summary: This diff introduces a new api for the stats functions. Previously, counter objects were stored as LockAndStatItem objects which were an std::pair of the lock and stat item. In order to facilitate the transition from using an std::pair to using folly/Synchronized for all locking purposes, a new LockableStat class is introduced, which is a simple wrapper around the previous LockAndStatItem object. This new wrapper class allows for addValue methods to be directly called on it and abstracts out all of the locking which previously had to be done by the user (user previously needed to first hold the spin lock, then call addValue). This change will help facilitate a smooth transition to replacing the LockAndStatItem object and updating the ExportedTimeseries class which uses it. Ultimately, the goal is to clean up many of the common/stats classes so that they can be moved to folly.

Differential Revision: D3458800

fbshipit-source-id: 8e618805bba926f15e8ad3ba63503326a3842b2a