extend HwSwitch state dump
Summary:
Add dumping of route table and interface table to
hw switch dump. We don't use it during warmboot, but rather to
enable on demand dumping of the hw switch state for debugging or
crafting bug repros. The data dumped is intended to try to capture
what is relevant to programming the switch hardware, or information
that comes back from the switch hardware about what we have
successfully programmed rather than being a dup of what we already
have from the SwitchState portion.

Differential Revision: D3349188

fbshipit-source-id: 9c6f0ba7e2daa6f8748ec106602cdf295a80e948