replace LockAndHistogram with new LockableHistogram class in fboss/agent directory
Summary: This diff introduces a new api for the stats functions. Previously, histograms were stored as LockAndHistogram objects which were an std::pair of the lock and histogram item. In order to facilitate the transition from using an std::pair to using folly/Synchronized for all locking purposes, a new LockableHistogram class is introduced, which is a simple wrapper around the previous LockAndHistogram object. This new wrapper class allows for addValue methods to be directly called on it and abstracts out all of the locking which previously had to be done by the user (user previously needed to first hold the spin lock, then call addValue). This change will help facilitate a smooth transition to replacing the LockAndHistogram object and updating the ExportedHistogramMap class which uses it. Ultimately, the goal is to clean up many of the common/stats classes so that they can be moved to folly.

Reviewed By: simpkins

Differential Revision: D3525120

fbshipit-source-id: 2ac807203873fe9a67ad223feac74dd998a806dd