Rename getStatPtr as getLockedStatPtr
Summary:
Previously, getStatPtr returned a LockedStatPtr object, while getUnlockedStatPtr returned a StatPtr object - this code cleans up the names of these methods such that getStatPtr returns a StatPtr and getLockedStatPtr returns a LockedStatPtr, as expected. This diff updates all call sites of getStatPtr to reflect this change.

The main distinction between these two methods is that getLockedStatPtr calls flush() on the timeseries object while getStatPtr does not. Though getLockedStatPtr is not explicitely needed, not having it would require each of these call sites to get the StatPtr, lock it, and then call flush on it before using it. The other alternative, having getStatPtr call flush before returning, is expensive and unecessary in cases where the user doesn't need any of the data, so this is not ideal either.

Reviewed By: yfeldblum

Differential Revision: D3584897

fbshipit-source-id: 0eb04f1cc65f519633ee23a70f62a58046772ada