set Tomahawk PCIe pre-emphasis for wedge100
Summary:
This properly sets the pcie pre-emphasis values on the
tomahawk for the wedge100 board. These values were determined by the
HW team in testing.

This is a little unfortunate that we need to make sdk calls from the
platform, but since the logic is platform specific I think it still
belongs in the platform code.

Differential Revision: D3631328

fbshipit-source-id: a0f44a62636a34d62940bb1ec12b0160bdd6807f