fix race condition in linkscan callback
Summary:
more details in the task, but essentially the problem is that if
a link flaps while we are adding or removing egresses, we will have
two threads simultaneously use the same datastructure (BcmHost::ecmpHosts_)
at the same time in an unsafe way. This seems particularly likely
during warm boots when we are re-creating a lot of egresses. The solution
is to make the linkscan callback code path get the egress ids from
the ASIC instead of looking at our cached state.

Differential Revision: D3623423

fbshipit-source-id: c5ef69b40f621e3b91a271e31b6aa5e662a87d29