Pull out Base class for ExportedStatMap
Summary: Pull out a parent class 'ExportedStatMapBase' that contains a cleaned up version of ExportedStatMap. This will make it easier to remove dependencies so that it can be moved to an open-source directory. ExportedStatMap now inherits from this new base class and will be kept around for the sake of backwards compatibility.

Reviewed By: yfeldblum

Differential Revision: D3646306

fbshipit-source-id: 0086b8acd997e05392e7cd68d075f142615e5bdf