fix deadlock from linkscan. send switch state update about port status
Summary:
Many details in the task, but tl;dr: calling portDown() directly
in the linkscan callback causes a deadlock where the cob is holding
a lock in the bcm sdk and wants the neighbor cache lock,
the update thread is holding our bcmswitch lock and wants the bcm sdk lock,
and the background thread is holding the neighbor cache lock and wants the
bcmswitch lock. Fix this issue by sending portDown() off to the backgound
evb. In addition, we want to refactor everything else out of the callback
to avoid this risk in the future for other calls, so add a switch state
update that signals that the port is down or up.

Reviewed By: aeckert

Differential Revision: D3645956

fbshipit-source-id: fa6b42698097c78d91b2f7a7f6e5ae35a6ac6e93