Pull out parent class for ExportedHistogramMap
Summary: Pull out a parent class 'ExportedHistogramMap' that contains a cleaned up version of ExportedHistogram (now renamed ExportedHistogramMapImpl). This will make it easier to remove dependencies so that it can be moved to an open-source directory. ExportedHistogramMapImpl now inherits from this new base class and will be kept around for the sake of backwards compatibility.

Reviewed By: yfeldblum

Differential Revision: D3700482

fbshipit-source-id: 6dd4f6c37ec8d111dcea05a418c34ef013d65e0f