don't return EX_SOFTWARE when communicating with missing qsfp
Summary:
We were always returning EX_SOFTWARE when we tried to set
settings on a port without a qsfp plugged in. There is no way of good
way (without more work) of determining qsfp not present vs. error
communicating over cp2112 chip. The message that we failed is likely
enough and the fact that there wasn't really a software error makes
this a reasonable change to me.

The other option would be to know exactly which ports to set, or to
handle this error code in TOR, but that seems worse.

Differential Revision: D3734125

fbshipit-source-id: f9dd7eb4cf833da1c76c9fff4455ebb0da2e0813