Reflect switch interfaces into Linux
Summary:
My first ever non-trivial diff on FBOSS-Agent .. I am new to this codebase and
your comments are most welcome to help improve the code.

== Why this diff:
Later during last half we had a discussion with FBOSS team about bringing
Aquaman onto FBOSS/Wedge for bright future of our networking softwares and
leveraging Wedges on EBB. Aquaman is now the sole IGP routing protocol on EBB so
we need to make it work on FBOSS to unblock our path to EBB.

Earlier we had monolithic `front0` for performing all of network IO to/from the
microserver. All addresses/subnets are added on front0 interface. This seems to
be a short workaround to make applications running on microserver to interact
other services over network.

== Problems:
- Linux had no visibility of switch interface states (for e.g. if a switch
  interface goes down then only it's addresses are removed)
- No explicit send/recv to/from actual switch interfaces as everything is
  routed by TunManager. For e.g. how do we do multicast

== Advantages:
- Reflects switch interfaces and it's state into linux. User application can
  enforce traffic/ping onto certain specific interface (TODO: needs a followup
  diff for performing routing in Software as opposed to hardware).
- Opens up opportunity for send/recv of multicast traffic (still needs routing
  part fix)
- Unblocks a step for Aquaman on FBOSS/Wedges and FBOSS/Wedge on EBB

== High level changes
- Instead of creating interface per RouterID now we create host-interface per switch-Interface
- On Host/Linux side: We have one routing table per switch Interface to do source based routing.
- Host->Agent : packets are received over tun-interface socketFd
- Agent->Host: host-Interface is identified by RxPacket's VLAN id and it is written to tun-interface socketFd

== Known Issues
- We rely on assumption that no one else messes up with links, rules and routes
  installed by TunManager. If someone does then it can lead to broken stage.

Reviewed By: tfangit

Differential Revision: D3736030

fbshipit-source-id: e18e68f7e8de7ee1992fdf2dee2f6a6b69fcf703