Change qsfp settings based on port speed
Summary:
Provides functionality so that one can pass a PortSpeed and have the requisite CDR values set.
Also does some cleanup around locking - previously it was possible to call this public function without a lock - made the actual logic bits private so this isn't possible. Also got rid of the customizeTransceiver method in the base class - it's only used by the QsfpModule - not Sfp - if it needs to be generalised, it can be added back, but for now I feel this is simpler.

Reviewed By: aeckert

Differential Revision: D3724110

fbshipit-source-id: 2d8b73a10d6cbabe0224ce5a0fe682063172b889