Reflect interface status in Linux
Summary:
Status of all `front*` interfaces is set to `IFF_UP` in Linux all the time
regardless of actual status of Ports belonging to that switch interface.

`Set<Port>` belongs to a `Vlan` aka `Interface` (as they have one to one
mapping). If all ports within `Vlan` is down then we declare `Vlan` as down and
unmark `IFF_UP` flag on that appropriate interface. Discussed this with Alex,
Jasmeet and Petr in person and we decided to go with this approach.

This diff should be NO-OP for services running on fboss-server. However Aquaman
will leverage netlink callbacks to learn Vlan statuses which will lead to much
faster convergences in routing.

=== Notes
- When agent restarts the interface status is changed to DOWN and when agent
  comes back up the status is changed to UP again which is as per expectation.

Reviewed By: tfangit

Differential Revision: D3769408

fbshipit-source-id: da61a65f2160b6c4b43329bf38f205078d349587