Enable LinkLocal routing to/from Host
Summary:
This is core diff which enables routing of LinkLocal/Multicast packets to/from
LinuxHost. At a high level this is divided into two parts

=== Routing from Switch ASIC to Host
- We have link-local route installed in LPM table with action 'FORWARD:TO_CPU'
- Once link local packet reaches the SwSwitch it is handled in Ipv4/Ipv6 handler
- In existing workflow we try to find the destination interface on host based on
  the destination address on the packet among global set of al addresses.
- However if packet is link-local or multicast we should just forward it to the
  host-interface corresponding to VLAN on which it was received. Link-local
  address must be assigned to interface for forwarding

=== Routing from Host to Switch ASIC
Packets received from Host (Tun) interface is L3. The current approach adds L2
header with cpu-mac as src/dst addr and tag it with specific VLAN tag. This
triggers L3 processing in switch.

This diff adds specific case for LinkLocal and Multicast addresses. We prepare
the full L2 packet with expected Mac-Address and pass it to corresponding VLAN
for L2 switching on that vlan. (VLAN is derived from Tun-interface ID).

Reviewed By: tfangit

Differential Revision: D3802833

fbshipit-source-id: d60d6a1e187adc7a1c95f21c0c6ea0dee93c4819