Add a platform arg to wedge_qsfp_util to enable picking right I2c bus
Summary:
The logic to pick right I2C bus in wedge_qsfp_util just looks
at the ASIC. If T2 it concludes that we want the WedgeI2CBus else the
Wedge100I2CBus. However with Galaxy in the mix this no longer works,
since its ASIC is not T2, but Wedge100I2CBus is not the right choice
for it.

To fix this add a platform arg, so we can invoke wedge_qsfp_util
with --platform galaxy.

Ideally, we would just detect the platform and figure out the right
I2C bus to create. This is doable when we are running the main FB
image. The use case being tackled here though is of a switch
booted in  wedge ramdisk and TOR making calls to wedge_qsfp_util (to
update QSFP settings and enable uplink detection). In ramdisk we don't
have the luxury of fbwhoami and fruid.json, but we still want to
invoke wedge_qsfp_util.

TODO: Its annoying to have to say --platform galaxy while running
in the main FB image. If no platform is specified -try detecting
platform via fbwhoami first if possible.

Depends on D3852696

Differential Revision: D3854801

fbshipit-source-id: c498c84be3ca70203a3be53ced02f49015904a45