wedge100: agent in usbError crash loop during version 90 roll
Summary:
While rolling wedge_agent:90 to wedge100s we noticed that for many of them
agent went into a continuous crash loop with the following error:

Main.cpp:108] switch initialization failed: facebook::fboss::UsbError: device not found

Doing 'lsusb' on switches that had this error we noticed that 'lsusb' did
not list the expected CP2112 device (Cygnal Integrated Products, Inc.) which
is used to bridge to I2C bus.

Doing a CP2112::resetDevice() when this happened failed to fix the issue (we
tested with a special binary based off wedge_agent:90).

Catching and ignoring the write error does get us past the problem.  Agent
comes up and seems to be working fine (bgp neighbors are up and we are able
to zping to the servers).

TODO:
This does not solve the problem of accessing I2C from microserver.  Also,
LEDs will probably be in some random state (however it may be better than
having agent crash when we are in this state.)  We should check if there is
something we can do from BMC to fix the issue.  Also, we should have send
this state to ODS so that we can alert when this happens.

Reviewed By: tfangit

Differential Revision: D3892255

fbshipit-source-id: 3ad5ba2393d18b0cce98bfa3f4dfa191824a6a78