Correct qsfp setup logic to prevent initialisation race conditions
Summary: There's a potential race condition between the time in which a switch is detected and when it's cache of data is refreshed - these happen at different cadences. This means it's possible for the transceiver to have been detected (present_) but dirty_ to be true. When retrieving information, this change now checks for this, handles it appropriatelu, and only then attempts to read the cached data.

Differential Revision: D3904297

fbshipit-source-id: 21c42913d09c4e333ca5965e1ed4496a473c922c