Disable using BMC REST endpoint
Summary:
Lately we have been seeing a ton of CP2112 not being detected alerts.
We suspect that BMC rest endpoint is hammering the switch that eventually
makes the kernel not detect the device anymore.

Inorder to test it, for now we will disable this REST endpoint and use the
'old' way of userver resetting the device when needed.

Reviewed By: aeckert

Differential Revision: D3915019

fbshipit-source-id: 45b02b9468d27b7991e66ac5b252f571c39d2944