No QSFPs for 6pack FCs
Summary:
We were incorrectly setting num qsfps for FCs to 16 (i.e.
the same number as 6pack LCs and wedge 40). This is not catastrophic
but causes us to do a bunch of extra work like probing for transceivers
when none would ever show up. Plus any code that tries to check if a QSFP
object exists for the port would wrongly conclude that the object
exists for backplane ports

Reviewed By: saifhhasan

Differential Revision: D3927372

fbshipit-source-id: 1f9e859627974e18e78aca695290e5c7e49b8aed