TunManager - avoid ugly hack of casting VlanID to IntfID
Summary:
We were assuming that VlanID is same as IntfID while forwarding packets to Linux
Host. This is not a good approach.

Here we pass explicit ID to TunManager to tell which interface in Linux should
the packet emerge from. This allows us having loopback interfaces without any
actual VLAN associated with them.

Differential Revision: D3935214

fbshipit-source-id: 5b99b3d468543dff936726e7a74fea3782207159