Disable FEC for Galaxy backplane ports even when running at 100G
Summary:
Ideally FEC should only be ever set for 100G optical ports.
But there are NICs which always turn these on regardless
of whether they are using optics or not. So we had FEC
enabled for all 100G ports.
However for Galaxy backplane ports we can work w/o FEC
even when running at 100G. This avoids the 100ns latency
incurred on a FEC enabled port (400ns when it goes from
one LC to another). So to work around and a api whereby
platform ports can be asked whether they can work with
FEC disabled.

Reviewed By: saifhhasan

Differential Revision: D3927684

fbshipit-source-id: 668481706709a10e813dbee7f71707fc4703c62c