Introduce special loopback interfaces
Summary:
Had a quick discussion meeting with Jasmeet and Cooper in afternoon regarding
this.

== Why ?
As of now we are sticking loopback addresses in FBOSS to one of the VLAN's
interface. However if all ports within VLAN is down then it also brings loopback
down which defeats the purpose of loopback interface.

In Galaxy broke LC101 brought down all BGP sessions among all FCs and LCs which
bad :(

== How we solve it ?
We agreed on having a special loopback interface (associated with VLAN-4095 or
no VLAN at all) and has those loopback addresses on it. This interface will
always be declared as UP on Linux host and usable regardless of port-statuses.

Differential Revision: D3935367

fbshipit-source-id: 65684a4269af7bc4752979dbb88da9f138ec2947